controller renamed to observer
